Organize event types in own files
